feat: updating controllers and requests

+++ app/Http/Controllers/PlanController.php
@@ -62,6 +62,8 @@ class PlanController extends BaseController
 
     public function destroy(Plan $plan): Response
     {
+        $this->authorize("delete", $plan);
+
         $plan->delete();
         $this->flushResourceCache();
